pyproject.toml
[project]
name = "django-starter-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "celery>=5.6.2",
  "django>=6.0.2",
  "django-allauth>=65.14.0",
  "django-celery-beat>=2.8.1",
  "django-celery-results>=2.6.0",
  "django-csp>=4.0",
  "django-daisy>=2.0.7",
  "django-htmx>=1.27.0",
  "django-split-settings>=1.3.2",
  "django-storages[s3]>=1.14.6",
  "django-tailwind-cli>=4.5.1",
  "environs[django]>=14.5.0",
  "flower>=2.0.1",
  "gunicorn>=25.0.2",
  "psycopg[binary]>=3.3.2",
  "redis>=7.1.0",
  "sentry-sdk>=2.52.0",
  "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
  "django-debug-toolbar>=6.2.0",
  "django-zeal>=2.0.4",
  "djlint>=1.36.4",
  "factory-boy>=3.3.3",
  "pyrefly>=0.50.1",
  "pytest>=9.0.2",
  "pytest-django>=4.11.1",
  "pytest-playwright>=0.4.4",
  "ruff>=0.14.14",
]

[tool.uv]
override-dependencies = [
  "django-celery-beat >= 2.8.1",
  "django >= 6.0.2"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests", "apps"]
addopts = "--reuse-db"

[tool.ruff]
line-length = 120
exclude = ["migrations"]

[tool.ruff.format]
quote-style = "single"

[tool.pyrefly]
python_version = "3.13"
project_includes = ["apps", "config"]
project_excludes = ["apps/**/tests"]
